Convert pipeline from function to script
As I am not testing this pipeline it makes more sense to keep it as a
script which is more readable
